backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0
pydantic==2.6.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
redis==5.0.1
celery==5.3.6
structlog==24.1.0
sentry-sdk==1.40.0
pytest==8.0.0
httpx==0.26.0
geoalchemy2==0.14.3
psycopg2-binary==2.9.9 # Helper for synchronous ops if needed, though using asyncpg
shapely==2.0.2
requests==2.31.0 # For synchronous external calls if needed, though httpx is better
